add dry run

+++ dedupe.js
@@ -4,23 +4,24 @@ var crypto = require('crypto');
 var path =   require('path');
 var R =      require('ramda');
 
-var dir = process.argv[2];
-if (!dir) {
-  console.error('Usage: ' + process.argv[1] + ' directory');
+var argv = require('minimist')(process.argv.slice(2), {boolean: ['n']});
+if (argv._.length !== 1) {
+  console.error('Usage: ' + process.argv[1] + ' [-n] directory');
+  console.error("  -n   Dry run. Don't delete any files, print what would be deleted");
   process.exit(1);
 }
 
-function checksum(str, algorithm, encoding) {
+function checksum(data, algorithm, encoding) {
   return crypto
     .createHash(algorithm || 'md5')
-    .update(str, 'utf8')
+    .update(data, 'utf8')
     .digest(encoding || 'hex');
 }
 
+var dir = argv._[0];
 var dupes = {};
-var files = fs.readdirSync(dir);
-files.forEach(function(f) {
-  var filename = path.join(dir, f);
+fs.readdirSync(dir).forEach(function(file) {
+  var filename = path.join(dir, file);
   var data = fs.readFileSync(filename);
   var hash = checksum(data);
   if (!dupes[hash]) {
@@ -34,6 +35,8 @@ R.valuesIn(dupes).forEach(function(dupe) {
   console.log('keep   ' + keep);
   dupe.forEach(function(filename) {
     console.log('remove ' + filename);
-    fs.unlinkSync(filename);
+    if (!argv.n) {
+      fs.unlinkSync(filename);
+    }
   });
 });
